17-fix-rotate (#23)
I have made a change to the offset handling. Because a display only knows 4 orientations. An offset of +3 with an orientation of 3 would be 6 and that is not a reasonable value. Technically this is permissible but does not lead to a meaningful result. The solution is the modulo operator. In this case, negative values are no longer required. Whether you take -1 or 3 leads to the same result.

Unfortunately, it took a little longer because my Linux distribution has only now reached gnome 45.